feature: dollar-rain (#17)
* feature: dollar-rain

In this commit, we introduce a fun and engaging "dollar rain" effect that triggers whenever a user performs a translation. This visual enhancement adds falling dollar bill animations to the screen, creating an exciting user experience. The effect is implemented using an HTML5 canvas for smooth performance and is customizable in terms of the number of bills, size, and duration. Additionally, we have removed the old translator module to streamline our codebase. This feature aims to make the translation process more enjoyable and visually appealing.

* feat: added loading gif

* refactor: remove dead code

* fix: jungle theme

* refactor: minor changes